Changelog: No negative AC/half physical dmg for fire/cold/shock/acid
Due to how L's resistance patch restructured resistable damage, negative
AC and half physical damage resistance weren't being applied
consistently to elemental monster attacks, so this is the way I had to
fix that, but I forgot to note that in the changelog, so here it is.

I don't know if this was the right call; high AC beyond a certain point
is already barely useful and this only exacerbates that problem.  This
is definitely worth revisiting at some point.

This is meant to go with commit ab978a15c3 (eat/crowning-gained
intrinsics provide only partial resistance).