Body armor skill reduces weight of worn body armor
Body armor skill reduces the weight of worn body armor heavier than
leather armor (150wt) by a percentage of the difference: 25%/50%/75% for
basic/skilled/expert respectively.

This is loosely based on the body armor weight reduction that Knights
receive in UnNetHack 5, though here it's based on the new body armor
skill and is thus available to all characters that can advance the
skill, as well as scaling with skill level.