Stop movement at low HP at strained instead of stressed
Movement at low HP (less than 10, or 5 if polymorphed) is now stopped at
strained (the same encumbrance that causes HP loss on movement) instead
of stressed, which simplifies the relationship between encumbrance, HP
loss and stopped movement a bit:

 * Before
    - Stressed
        - cannot move with less than 10 HP (5 if polymorphed)
    - Strained
        - lose 1 HP every 30 turns if moving
        - lose 1 HP every 3 turns if attacking
 * After
    - Strained
        - cannot move with less than 10 HP (5 if polymorphed)
        - lose 1 HP every 30 turns if moving
        - lose 1 HP every 3 turns if attacking

In other words, you know you're at risk of not being able to move at low
HP if you lose HP while moving due to encumbrance.  This is a minor buff
for players.

This also cleans up the code that does this in the process.