Make moving on Plane of Air easier by rolling vs. dexterity
Moving on the Plane of Air now only requires a d20 roll versus
dexterity, making it much easier to move on for most characters,
compared to the previous 75% chance to fail to move from NetHack.

This is yet another measure to make levitation a less mandatory part of
the ascension kit.  It does however make gauntlets of dexterity even
more desirable than its alternatives, which I might address later on.