Disable netclient by default, remove Jansson from build instructions
The netclient and server were features that were started but never
further developed by daniel_t for NitroHack.  Even NetHack Fourk (a
recent NetHack variant based on NetHack 4, which traces its roots back
to NitroHack) is hosted on the `nethack.xd.cm` public server behind
dgamelaunch instead.

The biggest problems with NitroHack's netclient and server features were
sluggishness and lack of security auditing (or being written with
security in mind to begin with).

Thus, I'm disabling it by default for DynaHack.  This also lets me drop
the Jansson requirement, which is an uncommon library on a lot of
platforms and was really awkward to have as a requirement, thus making
the build instructions easier to follow.