Add guaranteed tinning kit to Valkyrie quest home
If players can just spend thousands of turns eating giant corpses for
strength and waiting out the satiation outside of the quest portal, they
may as well get a tinning kit to speed up the process.  Case in point:
One of my own valkyries that got to the quest without a tinning kit did
exactly this and I quit the character because it was too dull.