Amulet of Yendor can map portals, gives more hints
Trying to find magic portals with the Amulet of Yendor under the rules
of NetHack 3.4.3 is a complete waste of time, i.e. wastes hundreds of
turns only to find that the portal was nowhere near where you were
looking.

This commit allows the Amulet of Yendor to straight up map detect a
magic portal if the player is near enough to it: the chance of this
happening is 2 divided by steps away from portal, which is polled every
1-in-steps-away actions (which in turn always gives the portal proximity
hint message).  The mapping itself is identical to reading a scroll of
gold detection while confused, but targeted only at the magic portal.

The net result: The closer you are to a magic portal with the Amulet of
Yendor in your inventory, the more likely you'll get a hint, which in
turn gives a higher chance of that portal being mapped.  Just going near
a magic portal has a good chance of mapping it.