Unify BUC state of shield of reflection in Perseus statue on Medusa
This removes the special casing for the chance of uncursed/cursed state
of the shield of reflection inside the statue of Perseus on Medusa's
Island, instead falling back on standard BUC chances that every other
object in the game gets.

It was weirdly inconsistent anyway: Titan Medusa had a 25% chance of
having a cursed shield, but all the other versions of the map had a 75%
chance instead.  This commit does away with all of this entirely, which
means one less thing to memorize.