Redesign of the Ranger quest home level (from dNetHack, by Khor)
The old layout with a long single-wide corridor and lots of wandering
quest friendlies and the occasional peaceful monster made it one of the
most annoying levels for players to navigate.

Obviously, DynaHack's level description syntax is a bit different from
dNetHack's, so a bit of adaption was needed to make this work.