Use rpath instead of `LD_LIBRARY_PATH` on *nix OSes
The rpath for the game binary can be set by configuring `GAME_RPATH` in
the cmake command, or after-the-fact using a utility like ccmake.  The
rpath itself is added upon invoking `make install`.

Internally, there's a bunch of machinery to support auto-setting
`GAME_RPATH` to use a relative path from the binary to
`libnitrohack.so`, but:

 1. `GAME_RPATH` is cached (which it must be to show up in the ccmake
    config GUI), so this auto-setting only occurs on the *first*
    invocation of cmake.

 2. Configuring paths on the initial cmake invocation is flaky due to
    some NitroHack-isms that I'll get around to fixing.

 3. The rpath is made relative using `$ORIGIN`, but this is only
    supported in Linux and Solaris, and notably *not* in Mac OS X, which
    instead uses @rpath / @loader_path / @executable_path in different
    circumstances I couldn't decipher.