Fix "bad fruit" warnings
When bones are saved, the fruit chain (i.e. mapping of fruit IDs to
fruit names that determines exactly what each slime mold is named) is
given negative IDs with the intent of flagging only the ones that will
be present in bones positive; a basic form of fruit-name garbage
collection that occurs when saving bones.

While fruit IDs are negative, any attempt to generate a name for a fruit
will fail, since the fruit-naming expects positive IDs.  Fruit naming
can occur when DynaHack's inventory sidebar is updated, whether or not
it's actually visible.

So... if the player dies, has a fruit of any kind in their inventory and
leaves bones, they get spammed with "Warning: Bad fruit #x?"

This commit fixes this issue by making naming more robust for fruits,
treating negative fruit IDs in the fruit chain as if they were positive.