Suppress an implicit type conversion warning in `options.c`
No change in logic, but suppresses a warning raised by clang.

The idea of the game interface code having to 'extend' the option types
with its own enum values like this is not great design, but that's how
daniel_t wrote the code, and redesigning it would require writing new
code, which runs the risk of introducing bugs.