Calculate sensible values for skill training > 100%
Now, if a basic skill is listed with "150%" training, it means that it
can be advanced to skilled and is 50% of the way to being advanced to
expert, instead of whatever the hell it meant before.

This also caps the percentages shown for skill training, so no more 567%
training next to frequently used skills.