Remove traps from Nymph level
Seriously, I enter the Nymph level with an elven cloak on for stealth,
put a towel over my eyes to use my telepathy, and not five turns later
nymphs are stepping on squeaky board traps and half of the nymphs are
already awake.  Let's make this a little more fair.