Replace Ranger quest goal iron bars with secret door
In dNetHack, iron bars can be negotiated with a wand of striking or
force bolt.  In DynaHack (and UnNetHack), you would need a potion of
acid instead, so replace iron bars blocking the alcove of trees with an
(obvious) locked secret door that mimics the stairs down on the Ranger
quest home level.