Prevent infinite looping for disconnected player input
Not sure why the hangup signal handler isn't doing this, but kerio has
run into 100% CPU DynaHack processes on nethack.xd.cm, so it clearly
wasn't doing its job.

Get rid of the vestiges of the 'blink' option while we're at it, since
the solution to this issue needs it totally gone in order to work.