ncurses: fix 'dark gray' rendering as solid black
I found an orcish helmet, which should have rendered as a gray '[', but instead
rendered as a solid black box. It appears that the custom color setup for
CURSES_DARK_GRAY is borked somehow, so just do what the TTY version does: use
bold black, which creates a grayish color instead.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>