add new MSGTYPE option
Add a new configuration option MSGTYPE

This allows the user to configure whether a message always waits for --more--,
or if it's never shown, or if it's never shown more than once in row.

The format is

    MSGTYPE=foo "string to match"

Where "foo" is one of "hide", "stop", "norep" or "show", and the string to
match is a message to be matched, with '*' and '?' wildcard expansion.

    "hide" hides the message, so it's never shown.
    "stop" always prompts for --more-- after the message.
    "norep" shows the message, but not again if no other message is shown in between.
    "show" always shows the message.

    ("noshow" is an alias to "hide", and "more" is an alias to "stop")

For example:

    # displacing pets
    MSGTYPE=hide "You displaced *"
    # items on floor
    MSGTYPE=norep "You see here a *"
    MSGTYPE=norep "You see here an *"
    # delayed instadeaths
    MSGTYPE=stop "You are slowing down."
    MSGTYPE=stop "You find it hard to breathe."
    MSGTYPE=stop "You are turning a little *"

Imported from https://bilious.alt.org/?397

Signed-off-by: Pasi Kallinen <paxed@alt.org>
Signed-off-by: Steven Noonan <steven@uplinklabs.net>